// src/types/courses/IRatingDet.ts
import IReview from "@itype/courses/IReview";

export default interface IRatingDet {
    r_1:      string;
    r_2:      string;
    r_3:      string;
    r_4:      string;
    r_5:      string;
    rc_1:      string;
    rc_2:      string;
    rc_3:      string;
    rc_4:      string;
    rc_5:      string;
}